fix: changing uint8_t to uint16_t from ec setup

+++ src/EC.h
@@ -7,9 +7,9 @@
 #include <util/delay.h>
 
 // Constantes para a leitura dos dados do EC11E
-#define __EC_READING_TIME               5000        // Tempo de leitura do rotary encoder (5ms)
-#define __EC_PULSES_PER_ROTATION		15          // Pulsos por rota??o do rotary encoder 
+#define __EC_READING_TIME               5000        // Tempo de leitura do rotary encoder (5000ms)
+#define __EC_PULSES_PER_ROTATION		15          // Pulsos por rotação do rotary encoder 
 #define __ANEMOMETER_RADIUS             100         // Raio do anemômetro (Tamanho das p?s): 100 mm
 
 void init_EC();                                     // Iniciar o EC11E;
-uint8_t Receive_data_EC();                          // Receber os dados do EC11E;
+uint16_t Receive_data_EC();                          // Receber os dados do EC11E;
